# backend/api/openapi/bingo.yaml
openapi: 3.1.0
info:
  title: Bingo Platform API
  version: 1.0.0
  description: |
    REST API for the Bingo Platform - Real-time multiplayer Bingo system
    with server-authoritative logic, supporting 1000+ concurrent players.

    **Key Features:**
    - Server-authoritative game logic
    - Cryptographically secure card generation
    - Real-time claim validation
    - Comprehensive penalty system
    - Audit logging and compliance

    **Performance Targets:**
    - Draw → UI propagation: <200ms average
    - Claim validation: <100ms server-side
    - Player join: <500ms from PIN to game entry
  contact:
    name: Bingo Platform Team
    email: support@bingo-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://bingo.local:3000
    description: Offline/LAN deployment
  - url: https://api.bingo-platform.com
    description: Production server
paths:
...
  /games:
    post:
      summary: Create a new game
      description: |
        Creates a new Bingo game with specified configuration.
        Generates unique 6-digit PIN and cryptographic game seed.
      operationId: createGame
      tags: [Games]
      security:
        - gamemaster: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    get:
      summary: List games
      description: List games with optional filtering
      operationId: listGames
      tags: [Games]
      security:
        - gamemaster: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/GameStatus'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Games retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameListResponse'
...
  /games/{gameId}/open:
    post:
      summary: Open game for players
      description: |
        Transitions game from LOBBY to OPEN status.
        Players can now join the game.
      operationId: openGame
      tags: [Games]
      security:
        - gamemaster: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Game opened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '400':
          description: Game cannot be opened (invalid state)
        '404':
          $ref: '#/components/responses/NotFound'
...
  /games/{gameId}/pause:
    post:
      summary: Pause the game
      description: Temporarily pause an active game
      operationId: pauseGame
      tags: [Games]
      security:
        - gamemaster: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Game paused successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
  /games/{gameId}/draw:
    post:
      summary: Draw next number
      description: |
        Manually draw the next number in sequence.
        Uses cryptographically secure RNG with game seed.
      operationId: drawNumber
      tags: [Games]
      security:
        - gamemaster: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      responses:
        '200':
          description: Number drawn successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Draw'
        '400':
          description: Cannot draw (game not active or complete)
  /games/{gameId}/auto-draw:
    post:
      summary: Toggle auto-draw
      description: Enable/disable automatic number drawing
      operationId: toggleAutoDraw
      tags: [Games]
      security:
        - gamemaster: []
      parameters:
        - $ref: '#/components/parameters/GameId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                enabled:
                  type: boolean
                interval:
                  type: integer
                  minimum: 5
                  maximum: 20
                  description: Auto-draw interval in seconds
              required: [enabled]
      responses:
        '200':
          description: Auto-draw settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
components:
  schemas:
...
    Game:
      type: object
      properties:
        id:
          type: string
          description: Unique game identifier
        pin:
          type: string
          pattern: '^[0-9]{6}$'
          description: 6-digit game PIN
        name:
          type: string
          maxLength: 100
          description: Optional game name
        status:
          $ref: '#/components/schemas/GameStatus'
        maxPlayers:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1000
        allowLateJoin:
          type: boolean
          default: true
        autoDrawInterval:
          type: integer
          minimum: 5
          maximum: 20
          default: 8
          description: Auto-draw interval in seconds
        winnerLimit:
          type: integer
          minimum: 1
          default: 1
        currentSequence:
          type: integer
          minimum: 0
        lastDrawAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          description: GameMaster identifier
      required:
        - id
        - pin
        - status
        - maxPlayers
        - allowLateJoin
        - autoDrawInterval
        - winnerLimit
        - currentSequence
        - createdAt
    GameStatus:
      type: string
      enum:
        - LOBBY
        - OPEN
        - ACTIVE
        - PAUSED
        - COMPLETED
        - CANCELLED
      description: |
        Game lifecycle status:
        - LOBBY: Being configured
        - OPEN: Accepting players
        - ACTIVE: Game in progress
        - PAUSED: Temporarily stopped
        - COMPLETED: Game finished
        - CANCELLED: Game aborted
...
    Draw:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        sequence:
          type: integer
          description: Draw order within game
        letter:
          type: string
          enum: ["B", "I", "N", "G", "O"]
        number:
          type: integer
          minimum: 1
          maximum: 75
        drawnAt:
          type: string
          format: date-time
        drawnBy:
          type: string
          description: GameMaster ID or 'AUTO'
      required:
        - id
        - gameId
        - sequence
        - letter
        - number
        - drawnAt
...
    CreateGameRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        maxPlayers:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1000
        allowLateJoin:
          type: boolean
          default: true
        autoDrawInterval:
          type: integer
          minimum: 5
          maximum: 20
          default: 8
        winnerLimit:
          type: integer
          minimum: 1
          default: 1
...
    GameListResponse:
      type: object
      properties:
        games:
          type: array
          items:
            $ref: '#/components/schemas/Game'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
        offset:
          type: integer
      required:
        - games
        - total
        - limit
        - offset
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error context
          nullable: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          description: Request tracking ID
      required:
        - error
        - message
        - timestamp
        - requestId
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            validationErrors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
                required:
                  - field
                  - code
                  - message
          required:
            - validationErrors
  parameters:
    GameId:
      name: gameId
      in: path
      required: true
      schema:
        type: string
      description: Unique game identifier
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: UNAUTHORIZED
            message: Invalid or missing JWT token
            timestamp: '2025-09-19T10:30:00Z'
            requestId: req_123456789
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NOT_FOUND
            message: Game not found
            timestamp: '2025-09-19T10:30:00Z'
            requestId: req_123456789
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Rate limit ceiling
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests in window
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: RATE_LIMITED
            message: Too many requests - please slow down
            details:
              retryAfter: 60
              limit: 100
              remaining: 0
            timestamp: '2025-09-19T10:30:00Z'
            requestId: req_123456789
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from joining a game.
        Token contains player-scoped claims and expires after 30 minutes.

        Example header: `Authorization: Bearer eyJhbGciOiJIUzI1NiIs...`

    gamemaster:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token with GameMaster privileges.
        Required for administrative operations.

        Example header: `Authorization: Bearer eyJhbGciOiJIUzI1NiIs...`
security:
  - jwt: []
...